+++ lib/cute_print/core_ext.rb
@@ -1,1 +1,2 @@
+require_relative "core_ext/irb"
 require_relative "core_ext/object"