+++ spec/support/captures_stderr.rb
@@ -1,5 +1,3 @@
-require_relative "../../test_support/captures_stderr"
-
 RSpec.configure do |c|
   c.include CapturesStderr
 end